feat: test for similar descriptions
