+++ package.json
@@ -20,6 +20,7 @@
     "@types/chai": "4.3.5",
     "@types/mocha": "10.0.1",
     "@types/node": "20.1.0",
+    "@types/string-similarity": "4.0.0",
     "@types/xml2js": "0.4.11",
     "chai": "4.3.7",
     "eslint": "8.40.0",
@@ -28,5 +29,8 @@
     "ts-mocha": "10.0.0",
     "typescript": "5.0.4",
     "xml2js": "0.5.0"
+  },
+  "dependencies": {
+    "string-similarity": "4.0.4"
   }
 }
